Day 24
Began with a basically brute-force approach: only the value
of Z matters on entry to each "inp" segment; we record the
resulting value of Z on output together with the largest
(resp, smallest) set of digits that gave that output.

Because many prefixes of the inputs produce the same values of Z
when they're consumed, we can fold all of those intermediate
stages down.

The alternative version brings Z3 to bear. This really struggles
using the in-built factor of 26; it struggles using "unbounded"
ints as opposed to fixed-length bit vectors, also. Optimising
away both of those (only really doable with some inspection),
Z3 can produce answers almost instantaneously.